semantics: move all diagnostic logic into `src/diagnostics`
