+++ compiler/hash-ast-passes/src/diagnostics/error.rs
@@ -1,7 +1,6 @@
-//! Hash AST semantic passes error definitions.
+//! Hash AST semantic analysis error diagnostic definitions.
 //!
 //! All rights reserved 2022 (c) The Hash Language authors.
-use std::fmt::Display;
 
 use hash_ast::ast::Visibility;
 use hash_error_codes::error_codes::HashErrorCode;
@@ -9,6 +8,8 @@ use hash_reporting::reporting::{
     Report, ReportBuilder, ReportCodeBlock, ReportElement, ReportKind, ReportNote, ReportNoteKind,
 };
 use hash_source::location::SourceLocation;
+
+use super::{BlockOrigin, PatternOrigin};
 
 /// An error that can occur during the semantic pass
 pub struct AnalysisError {
@@ -60,73 +61,6 @@ pub(crate) enum AnalysisErrorKind {
         modifier: Visibility,
         origin: BlockOrigin,
     },
-}
-
-/// Denotes where a pattern was used as in the parent of the pattern. This is useful
-/// for propagating errors upwards by signalling what is the current parent of the
-/// pattern. This only contains patterns that can be compound (hold multiple children patterns).
-#[derive(Clone, Copy, Debug)]
-pub(crate) enum PatternOrigin {
-    Tuple,
-    NamedField,
-    Constructor,
-    List,
-    Namespace,
-}
-
-impl PatternOrigin {
-    /// Convert the [PatternOrigin] into a string which can be used for displaying
-    /// within error messages.
-    fn to_str(self) -> &'static str {
-        match self {
-            PatternOrigin::Tuple => "tuple",
-            PatternOrigin::NamedField => "named field",
-            PatternOrigin::Constructor => "constructor",
-            PatternOrigin::List => "list",
-            PatternOrigin::Namespace => "namespace",
-        }
-    }
-}
-
-impl Display for PatternOrigin {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.to_str())
-    }
-}
-
-/// Denotes where an error occurred from which type of block. This is useful
-/// when giving more context about errors such as [AnalysisErrorKind::NonDeclarativeExpression]
-/// occur from.
-#[derive(Clone, Copy, Debug)]
-pub(crate) enum BlockOrigin {
-    Root,
-    Mod,
-    Impl,
-    Body,
-}
-
-impl BlockOrigin {
-    /// Convert the [BlockOrigin] into a string which can be used for displaying
-    /// within error messages.
-    fn to_str(self) -> &'static str {
-        match self {
-            BlockOrigin::Root | BlockOrigin::Mod => "module",
-            BlockOrigin::Impl => "impl",
-            BlockOrigin::Body => "body",
-        }
-    }
-}
-
-impl Default for BlockOrigin {
-    fn default() -> Self {
-        BlockOrigin::Root
-    }
-}
-
-impl Display for BlockOrigin {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.to_str())
-    }
 }
 
 impl From<AnalysisError> for Report {
